refactor: moved views into their respective component folders
