+++ src/components/internshipList/InternshipListView.vue
