+++ src/components/viewInternship/ViewInternshipView.vue
